# chess_backend/interfaces/openapi.running.json
{"openapi":"3.1.0","info":{"title":"Chess Backend API","description":"Backend service for a fullstack chess application. Provides user management, game creation, move validation (legal chess rules), persistence with SQLite, and explicit save/load via named snapshots.","version":"0.2.0"},"paths":{"/users":{"post":{"tags":["users"],"summary":"Create a user","description":"Create a new user with a unique username.","operationId":"create_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"tags":["users"],"summary":"Get a user by ID","description":"Fetch a user record by user_id.","operationId":"get_user","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games":{"post":{"tags":["games"],"summary":"Create a new game","description":"Create a new chess game. Players are optional; if provided, they must exist.","operationId":"create_game","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGameRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Bad Request"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["games"],"summary":"List games (optionally filtered by user)","description":"List games. If userId is provided, returns games where the user is white or black.","operationId":"list_games","parameters":[{"name":"userId","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter games by participant userId.","title":"Userid"},"description":"Filter games by participant userId."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GameResponse"},"title":"Response List Games"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}":{"get":{"tags":["games"],"summary":"Get a game by ID","description":"Fetch a game including current state and move history.","operationId":"get_game","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"string","title":"Game Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["games"],"summary":"Delete a game (optional)","description":"Delete a game and its moves/snapshots.","operationId":"delete_game","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"string","title":"Game Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}/moves":{"post":{"tags":["games"],"summary":"Submit a move","description":"Submit a chess move (SAN or UCI). The move is validated and applied if legal.","operationId":"submit_move","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"string","title":"Game Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitMoveRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}/save":{"post":{"tags":["games"],"summary":"Save a named snapshot","description":"Save the current game state (FEN + move_count) as a named snapshot. Upserts by name.","operationId":"save_game_snapshot","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"string","title":"Game Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveGameRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SnapshotResponse"}}}},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/games/{game_id}/load":{"post":{"tags":["games"],"summary":"Load a named snapshot","description":"Restore game state from a named snapshot. Moves after snapshot move_count are discarded.","operationId":"load_game_snapshot","parameters":[{"name":"game_id","in":"path","required":true,"schema":{"type":"string","title":"Game Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoadGameRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GameResponse"}}}},"400":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Bad Request"},"404":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["health"],"summary":"Health check","description":"Simple health check endpoint used by deployment and monitoring.","operationId":"health_check","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CreateGameRequest":{"properties":{"white_user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"White User Id","description":"User ID for white player (optional)."},"black_user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Black User Id","description":"User ID for black player (optional)."}},"type":"object","title":"CreateGameRequest"},"CreateUserRequest":{"properties":{"username":{"type":"string","maxLength":64,"minLength":1,"title":"Username","description":"Unique username."}},"type":"object","required":["username"],"title":"CreateUserRequest"},"ErrorResponse":{"properties":{"detail":{"type":"string","title":"Detail","description":"Human-readable error message."}},"type":"object","required":["detail"],"title":"ErrorResponse"},"GameResponse":{"properties":{"id":{"type":"string","title":"Id","description":"Game ID (UUID)."},"white_user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"White User Id","description":"User ID for white player."},"black_user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Black User Id","description":"User ID for black player."},"fen":{"type":"string","title":"Fen","description":"Current board state as FEN."},"turn":{"type":"string","title":"Turn","description":"Side to move: 'w' or 'b'."},"status":{"type":"string","title":"Status","description":"Game status: active/checkmate/stalemate/draw."},"is_check":{"type":"boolean","title":"Is Check","description":"Whether the side to move is currently in check."},"moves":{"items":{"$ref":"#/components/schemas/MoveResponse"},"type":"array","title":"Moves","description":"Move history."}},"type":"object","required":["id","fen","turn","status","is_check"],"title":"GameResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoadGameRequest":{"properties":{"name":{"type":"string","maxLength":64,"minLength":1,"title":"Name","description":"Snapshot name to restore.","default":"latest"}},"type":"object","title":"LoadGameRequest"},"MoveResponse":{"properties":{"id":{"type":"string","title":"Id","description":"Move ID (UUID)."},"move_number":{"type":"integer","title":"Move Number","description":"Sequential move number starting at 1."},"uci":{"type":"string","title":"Uci","description":"Move in UCI format, e.g. e2e4, g1f3, e7e8q."},"san":{"type":"string","title":"San","description":"Move in SAN format, e.g. e4, Nf3, O-O."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"UTC timestamp when the move was created."}},"type":"object","required":["id","move_number","uci","san","created_at"],"title":"MoveResponse"},"SaveGameRequest":{"properties":{"name":{"type":"string","maxLength":64,"minLength":1,"title":"Name","description":"Snapshot name.","default":"latest"}},"type":"object","title":"SaveGameRequest"},"SnapshotResponse":{"properties":{"id":{"type":"string","title":"Id","description":"Snapshot ID."},"game_id":{"type":"string","title":"Game Id","description":"Associated game ID."},"name":{"type":"string","title":"Name","description":"Snapshot name."},"fen":{"type":"string","title":"Fen","description":"FEN stored in snapshot."},"move_count":{"type":"integer","title":"Move Count","description":"Move count when snapshot was taken."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"UTC timestamp when snapshot was created."}},"type":"object","required":["id","game_id","name","fen","move_count","created_at"],"title":"SnapshotResponse"},"SubmitMoveRequest":{"properties":{"uci":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Uci","description":"Move in UCI format (preferred for programmatic clients)."},"san":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"San","description":"Move in SAN format (common in chess notation)."}},"type":"object","title":"SubmitMoveRequest"},"UserResponse":{"properties":{"id":{"type":"string","title":"Id","description":"User ID (UUID)."},"username":{"type":"string","title":"Username","description":"Unique username."},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"UTC timestamp when the user was created."}},"type":"object","required":["id","username","created_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"health","description":"Health and service info endpoints."},{"name":"users","description":"User management (create and fetch users)."},{"name":"games","description":"Chess game management (create games, play moves, save/load snapshots)."}]}